Split distributed test into spatial and nearest
It was too unwieldy.
